Create the top 10 AWS account security items recommended on the AWS Security blog
The AWS Security Blog post Top 10 security items to improve in your AWS account" contains tips for improving your cloud security contains helpful tips to help you improve security in your AWS accounts.

In Build a custom benchmark for the top 10 AWS security tips we show one approach to mapping the items from the blog post to Steampipe controls. This benchmark can be used as an example that we encourage you to reuse and remix based on your security requirements.